+++ src/templates/signup-block/components/social-signup.vue
@@ -1,0 +1,33 @@
+<template>
+  <div class="flex flex-col gap-4">
+    <PrimeButton
+      label="Continue with Google"
+      icon="pi pi-google"
+      outlined
+    />
+    <PrimeButton
+      label="Continue with Github"
+      icon="pi pi-github"
+      outlined
+    />
+  </div>
+  <div class="w-full">
+    <Divider
+      layout="horizontal"
+      align="center"
+      >or</Divider
+    >
+  </div>
+  <PrimeButton
+    label="Sign Up with E-mail"
+    outlined
+    @click="$emit('show-form')"
+  />
+</template>
+
+<script setup>
+  import PrimeButton from 'primevue/button'
+  import Divider from 'primevue/divider'
+
+  defineEmits(['show-form'])
+</script>
